Проработал большую часть Code Review
Общее:
1. Провел reformat всех классов
2. Произвел оптимизацию импортов во всех классах
3.Создал util класс PaginationForm для уменьшения дублирвания кода. В нем реализовал метод обработки пагинации

Конфиги:
1. LoggerAspect удалил закоммиченный код, по надобности можно найти его в git

Контроллеры:
1. В CommentController в методах deleteCommentForPostIdAndCommentId, editCommentForPostIdAndCommentId,recoverDeletedCommentForPostIdAndCommentId параметр comment_id привел в порядок (camelCase)
2. Аналогично в GlobalExceptionHandlerController
3. В SearchController в именах параметров поправил snake_case
4. UserRegisterController вывел http в константу во избежание дупликатов кода

DTO:
1. В MessageEntity нейминг поля sent_time привел к camelCase

Entity:
1. имплементировал все Entity от Serializable
2. Вместо метода getLongFromEnum в классе FriendshipStatusType добавил в Enum переменную value, убрал switch. Везде где использовался switch-метод исправил все на использоввание value

Helper:
1. BeanInitMethodImpl убрал impl в имени

Mapper:
1. Удалил CommentMapStructService

Repository:
1. FriendshipRepository поменял нативный запрос на jpql

Services:
1. В PostService вынес обработку ошибок в отдельный метод

Security:
1. В классах CustomAccessDeniedHandler и CustomAuthenticationEntryPoint для response поменял установление статус-коды на константы
2. JwtBlackListService удалил конструктор, заменил на @RequiredArgsConstructor
3. В классе JWTUtilService в методах isTokenExpired и validateToken возврат примитива, а не объекта
4. В классе SocialNetUserDetails проверку userRoleEntityList.size() != 0 заменил на метод isEmpty()

+++ src/main/java/ru/example/group/main/repository/UserRepository.java
@@ -11,14 +11,17 @@ import java.util.List;
 @Repository
 public interface UserRepository extends JpaRepository<UserEntity, Long> {
     UserEntity findByEmail(String eMail);
+
     Boolean existsByEmail(String eMail);
+
     UserEntity findByConfirmationCode(String code);
+
     @Query(value = """
             SELECT users.* FROM users
             WHERE (users.is_approved = true AND users.is_blocked = false AND users.is_deleted = false) AND users.id
             IN (SELECT friendships.src_person_id FROM friendships WHERE friendships.dst_person_id = :id AND ((friendships.status_id) = :status))
             """
             , nativeQuery = true)
-    List<UserEntity> getAllRelationsOfUser(Long id, Long status, Pageable pageable);
+    List<UserEntity> getAllRelationsOfUser(Long id, int status, Pageable pageable);
 }
